requirements.txt
setuptools
labelImg
flask
ultralytics
numpy
opencv-python
pillow
torch
